Added the error flash whenever a campground is not found.

+++ routes/campgrounds.js
@@ -58,6 +58,10 @@ router.get(
 	catchAsync(async (req, res) => {
 		const { id } = req.params;
 		const ground = await Campground.findById(id);
+		if(!ground){
+			req.flash('error', 'Requested campground cannot be found');
+			return res.redirect('/campgrounds');
+		}
 		res.render("campground/edit", { ground, cities });
 	})
 );
@@ -86,6 +90,11 @@ router.delete(
 	"/:id",
 	catchAsync(async (req, res) => {
 		const { id } = req.params;
+		const ground = await Campground.findById(id);
+		if(!ground){
+			req.flash('error', 'Requested campground cannot be found');
+			return res.redirect('/campgrounds');
+		}
 		await Campground.findByIdAndDelete(id);
 		req.flash("success", "Successfully deleted the campground");
 		res.redirect("/campgrounds");
